+++ librecomp/include/librecomp/game.hpp
@@ -9,21 +9,12 @@
 #include <ultramodern/ultramodern.hpp>
 
 namespace recomp {
-    enum class SaveType {
-        None,
-        Eep4k,
-        Eep16k,
-        Sram,
-        Flashram,
-        AllowAll, // Allows all save types to work and reports eeprom size as 16kbit.
-    };
-
     struct GameEntry {
         uint64_t rom_hash;
         std::string internal_name;
         std::u8string game_id;
         std::string mod_game_id;
-        SaveType save_type = SaveType::None;
+        ultramodern::SaveType save_type = ultramodern::SaveType::None;
         bool is_enabled;
         // Only needed for mod function hooking support, not needed if `has_compressed_code` is false.
         std::vector<uint8_t> (*decompression_routine)(std::span<const uint8_t> compressed_rom) = nullptr;
@@ -109,7 +100,6 @@ namespace recomp {
     /// 
     void start(const Configuration& cfg);
 
-    SaveType get_save_type();
     bool eeprom_allowed();
     bool sram_allowed();
     bool flashram_allowed();
